% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_composer}
\alias{get_composer}
\title{Get list of compositions}
\usage{
get_composer(composer, file_ext = NULL)
}
\arguments{
\item{composer}{a character, composer or collection name, e.g. \code{bach}.}

\item{file_ext}{if provided, defines which extensions are returned.
\code{file_ext=NULL} (default) returns all extensions.}
}
\description{
Return all filenames in the corpus that match a composer's or a
collection's name.
}
\examples{
{
get_composer("schoenberg")
}
}
