% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{load_music}
\alias{load_music}
\title{Load music from work name}
\usage{
load_music(path)
}
\arguments{
\item{path}{character internal path, e.g. \code{bach/bwv66.6}}
}
\description{
Loads music from music21 internal path
}
