% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{music21}
\alias{music21}
\alias{music21-package}
\alias{music21}
\title{music21 for R}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
music21
}
\description{
\href{http://web.mit.edu/music21/doc/index.html}{music21} is an open source software library
for structured music data IO and manipulation.

python music21 complete module. Treat it like an RC object
}
\details{
The music21 package provides access to the
complete music21 API from within R.
}
\keyword{datasets}
