% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{plot.music21.base.Music21Object}
\alias{plot.music21.base.Music21Object}
\alias{view.music21.base.Music21Object}
\title{Shows music in the viewer pane}
\usage{
\method{plot}{music21.base.Music21Object}(x, ...)

\method{view}{music21.base.Music21Object}(x, ...)
}
\arguments{
\item{x}{music21 python object}

\item{...}{other options (currently ignored)}
}
\description{
This function uses \href{http://lilypond.org/}{lilypond} to save the music21
object to a png file, then loads the image.
}
