% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.music21.base.Music21Object}
\alias{print.music21.base.Music21Object}
\title{Prints music21 generic object}
\usage{
\method{print}{music21.base.Music21Object}(x, ...)
}
\arguments{
\item{x}{music21 python object}

\item{...}{other options (currently ignored)}
}
\description{
Uses show method from music21
}
