% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_music}
\alias{read_music}
\alias{read_midi}
\alias{read_musicxml}
\title{Load midi or musicxml music into a stream}
\usage{
read_music(file, ext = NULL)

read_midi(file)

read_musicxml(file)
}
\arguments{
\item{file}{path of the file}

\item{ext}{extension of the file. If NULL (default) gets the extension from
file name.}
}
\description{
Load midi or musicxml music into a stream object
}
