% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_lily}
\alias{write_lily}
\title{Write music to lily PNG file}
\usage{
write_lily(x, path = NULL)
}
\arguments{
\item{x}{music21 object}

\item{path}{path to save PNG file. If NULL (default) returns a tempfile.}
}
\description{
Write music21 to png file using lilypond
}
